refactor: move hexGrid, coordinateKey, serialization to core/utils
