+++ src/hooks/usePersistence.ts
@@ -1,6 +1,6 @@
 import { useEffect } from 'react'
 import { useStore } from '../store/useStore'
-import { encodePieces, decodePieces } from '../utils/serialization'
+import { encodePieces, decodePieces } from '../core/utils/serialization'
 
 const STORAGE_KEY = 'naval-planner-design'
