+++ tests/validationTri.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect } from 'vitest'
 import { canPlace, isTriInBounds } from '../src/utils/validation'
 import type { PlacedPiece, PiecesConfig } from '../src/types'
-import { toTriKey, toTriEdgeKey } from '../src/utils/coordinateKey'
+import { toTriKey, toTriEdgeKey } from '../src/core/utils/coordinateKey'
 import piecesConfig from '../src/data/pieces-config.json'
 
 const config = piecesConfig as PiecesConfig
